Few new learnsets, add more fossil revive options
Learnsets are not complete but are looking really
good now overall. Added more fossil revival options
to the Pewter museum. Not sure if what I added is
enough to make it work. readme update. Little tweaks
to evolution levels/methods and move data. Completed
reworks of Houndour and Phanpy lines, maybe a couple
of others.